+++ lib/rust/ensogl/component/scrollbar/src/lib.rs
@@ -27,14 +27,11 @@ use ensogl_core::application::Application;
 use ensogl_core::data::color;
 use ensogl_core::display;
 use ensogl_core::display::shape::StyleWatchFrp;
-use ensogl_core::timer::DelayedRepeats;
 use ensogl_core::Animation;
 use ensogl_hardcoded_theme as theme;
 use ensogl_selector as selector;
 use ensogl_selector::model::Model;
 use ensogl_selector::Bounds;
-
-use std::time::Duration;
 
 
 
@@ -56,10 +53,10 @@ pub const PADDING: f32 = 3.0;
 const MIN_THUMB_SIZE: f32 = 12.0;
 /// After an animation, the thumb will be visible for this time, before it hides again.
 const HIDE_DELAY: f32 = 1000.0;
-/// Time delay before holding the mouse button triggers scrolling.
-const HOLD_REPEAT_DELAY: Duration = Duration::from_millis(500);
-/// Time interval between scrolls when holding the mouse button.
-const HOLD_REPEAT_INTERVAL: Duration = Duration::from_millis(200);
+/// Time delay before holding down a mouse button triggers first scroll, in milliseconds.
+const CLICK_AND_HOLD_DELAY_MS: i32 = 500;
+/// Time interval between scrolls while holding down a mouse button, in milliseconds.
+const CLICK_AND_HOLD_INTERVAL_MS: i32 = 200;
 
 const ERROR_MARGIN_FOR_ACTIVITY_DETECTION: f32 = 0.1;
 
@@ -128,8 +125,11 @@ impl Frp {
         let bg_default_color = style.get_color(theme::component::slider::background::color);
         let bg_hover_color = style.get_color(theme::component::slider::background::hover_color);
 
-        let background_hold_repeats =
-            DelayedRepeats::new(network, HOLD_REPEAT_DELAY, HOLD_REPEAT_INTERVAL);
+        let click_and_hold_timer = frp::io::timer::DelayedInterval::new(
+            network,
+            CLICK_AND_HOLD_DELAY_MS,
+            CLICK_AND_HOLD_INTERVAL_MS,
+        );
 
         frp::extend! { network
 
@@ -246,21 +246,23 @@ impl Frp {
                 });
 
 
-            // === Click and hold repeats ===
+            // === Click and hold repeated scrolling ===
 
-            background_hold_repeats.active <+ base_frp.is_dragging_background;
-            background_hold <- mouse_position.sample(&background_hold_repeats.on_trigger);
+            click_and_hold_timer.set_active <+ base_frp.is_dragging_background;
+            mouse_pos_at_time_trigger <- mouse_position
+                .sample(&click_and_hold_timer.on_trigger);
 
-            thumb_scale <- all_with(&inner_length, &frp.set_max, |length_px, max| 1.0 / length_px * max);
-            frp.scroll_by <+ background_hold.map4(&thumb_center_px, &frp.set_thumb_size, &thumb_scale,
-                |hold_position, thumb_center, thumb_size, thumb_scale| {
-                    let hold_distance = (hold_position.x - *thumb_center) * thumb_scale;
-                    if hold_distance.abs() < thumb_size * 0.5 {
-                        0.0
-                    } else {
-                        let direction = hold_distance.signum();
-                        direction * thumb_size * CLICK_JUMP_PERCENTAGE
-                    }
+            offset_from_thumb_px <- mouse_pos_at_time_trigger.map2(&thumb_center_px,
+                |mouse_pos, thumb_center| thumb_center - mouse_pos.x);
+            offset_from_thumb <- offset_from_thumb_px.map3(&inner_length, &frp.set_max,
+                |offset_px, length_px, max| offset_px / length_px * max);
+
+            frp.scroll_by <+ offset_from_thumb.map2(&frp.set_thumb_size,
+                |mouse_offset, thumb_size| {
+                    let half_thumb_size = thumb_size / 2.0;
+                    let mouse_outside_thumb = mouse_offset.abs() > half_thumb_size;
+                    let direction = if mouse_outside_thumb { -mouse_offset.signum() } else { 0.0 };
+                    direction * thumb_size * CLICK_JUMP_PERCENTAGE
                 });